requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
django-appconf==1.0.5; python_version >= '3.6'
django-compressor==3.1
django-libsass==0.9
django-markdownx==4.0.0b1
django==4.0.3
importlib-metadata==4.11.3; python_version < '3.10'
libsass==0.21.0
markdown==3.3.6
pillow==9.0.1; python_version >= '3.7'
psycopg2-binary==2.9.3
rcssmin==1.1.0
rjsmin==1.2.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
zipp==3.7.0; python_version >= '3.7'
